#000: Mausam : Extract actorRef codec and CborAkkaSerializer into common module
