+++ csw-command/csw-command-client/src/main/scala/csw/command/client/cbor/CommandAkkaSerializer.scala
@@ -6,6 +6,7 @@ import akka.actor.typed.scaladsl.adapter._
 import csw.command.client.messages._
 import csw.command.client.messages.sequencer.CswSequencerMessage
 import csw.command.client.models.framework._
+import csw.commons.cbor.CborAkkaSerializer
 import csw.params.commands.CommandResponse
 import csw.params.core.states.StateVariable
 import csw.serializable.CommandSerializable
