+++ csw-commons/src/main/scala/csw/commons/cbor/CborAkkaSerializer.scala
@@ -1,4 +1,4 @@
-package csw.command.client.cbor
+package csw.commons.cbor
 
 import akka.serialization.Serializer
 import csw.logging.api.scaladsl.Logger
