+++ csw-commons/src/main/scala/csw/commons/codecs/ActorCodecs.scala
@@ -1,0 +1,17 @@
+package csw.commons.codecs
+
+import akka.actor.typed.{ActorRef, ActorRefResolver, ActorSystem}
+import io.bullet.borer.Codec
+
+trait ActorCodecs {
+  implicit def actorSystem: ActorSystem[_]
+
+  implicit def actorRefCodec[T]: Codec[ActorRef[T]] = {
+    val resolver = ActorRefResolver(actorSystem)
+
+    Codec.bimap[String, ActorRef[T]](
+      resolver.toSerializationFormat,
+      resolver.resolveActorRef
+    )
+  }
+}
